Tidy aarch64 registers
